Refactor: signup부분의 기능을 util.js파일에 따로모아 실행되도록 수정
